LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/navigate-from-initial-about-blank-gc-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: Invalid URL
CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: Invalid URL


PASS navigate() from <iframe> with src="" but still on initial about:blank doesn't cause a crash on GC
